# Faster overlapping copies