@nigeltao 2019-09-01
Code looks great. Again, though, I'd like to know what the GOARCH is that you measured, in the commit message's numbers.